File restructure using Cursor
