+++ app/src/main/java/com/example/template/ui/screens/exercises/ExercisesScreen.kt
@@ -1,0 +1,11 @@
+package com.example.template.ui.screens.exercises
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun ExercisesScreen() {
+	Text("Exercises screen (coming soon)")
+}
+
+
