+++ app/src/main/java/com/example/template/ui/screens/meals/MealsScreen.kt
@@ -1,0 +1,11 @@
+package com.example.template.ui.screens.meals
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun MealsScreen() {
+	Text("Meals screen (coming soon)")
+}
+
+
