+++ app/src/main/java/com/example/template/ui/theme/Type.kt
@@ -1,34 +1,84 @@
+@file:JvmName("TypeKt")
+
 package com.example.template.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
 import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
+import com.example.template.R // Your project's R file
 
-// Set of Material typography styles to start with
+// Define the Inter FontFamily using your variable fonts
+val InterFontFamily = FontFamily(
+    // Reference the upright variable font for normal style
+    Font(R.font.inter_variable, FontWeight.Normal, FontStyle.Normal), // Catches all normal weights
+
+    // Reference the italic variable font for italic style
+    Font(
+        R.font.inter_italic_variable,
+        FontWeight.Normal,
+        FontStyle.Italic
+    ) // Catches all italic weights
+    // Note: The FontWeight.Normal here is a placeholder for the variable font.
+    // The actual weight will be determined by the FontWeight you specify in your TextStyle.
+)
+
+// Now, update your Typography object to use InterFontFamily
+// and specify the desired FontWeight (and FontStyle if needed) for each text style.
 val Typography = Typography(
-    bodyLarge = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 16.sp,
-        lineHeight = 24.sp,
-        letterSpacing = 0.5.sp
-    )
-    /* Other default text styles to override
+    displayLarge = TextStyle(
+        fontFamily = InterFontFamily,
+        fontWeight = FontWeight.W400, // Example: Corresponds to Regular in Inter
+        fontSize = 57.sp,
+        lineHeight = 64.sp,
+        letterSpacing = (-0.25).sp
+    ),
+    // ... (other display styles) ...
+
+    headlineLarge = TextStyle(
+        fontFamily = InterFontFamily,
+        fontWeight = FontWeight.W600, // Example: Corresponds to SemiBold in Inter
+        fontSize = 32.sp,
+        lineHeight = 40.sp,
+        letterSpacing = 0.sp
+    ),
+    // ... (other headline styles) ...
+
     titleLarge = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
+        fontFamily = InterFontFamily,
+        fontWeight = FontWeight.W500, // Example: Corresponds to Medium in Inter
         fontSize = 22.sp,
         lineHeight = 28.sp,
         letterSpacing = 0.sp
     ),
+    // ... (other title styles) ...
+
+    bodyLarge = TextStyle(
+        fontFamily = InterFontFamily,
+        fontWeight = FontWeight.W400, // Regular
+        fontSize = 16.sp,
+        lineHeight = 24.sp,
+        letterSpacing = 0.5.sp
+    ),
+    bodyMedium = TextStyle(
+        fontFamily = InterFontFamily,
+        fontWeight = FontWeight.W400, // Regular
+        fontSize = 14.sp,
+        lineHeight = 20.sp,
+        letterSpacing = 0.25.sp,
+        // Example if you wanted this specific style to be italic:
+        // fontStyle = FontStyle.Italic
+    ),
+    // ... (other body and label styles, adjusting FontWeight and FontStyle as needed) ...
+
     labelSmall = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Medium,
+        fontFamily = InterFontFamily,
+        fontWeight = FontWeight.W500, // Medium
         fontSize = 11.sp,
         lineHeight = 16.sp,
         letterSpacing = 0.5.sp
     )
-    */
 )
